app/requirements.txt
praw
pymongo
requests
beautifulsoup4
logging
python-dotenv